+++ libs/expo/betterangels/src/lib/providers/index.ts
@@ -1,5 +1,6 @@
 export * from './featureControls';
 export { default as KeyboardToolbarProvider } from './keyboardToolbar/keyboardToolbarProvider';
+export { default as NativePaperProvider } from './nativePaper/NativePaperProvider';
 export { default as SnackbarProvider } from './snackbar/SnackbarProvider';
-export { default as UserProvider } from './user/UserProvider';
 export * from './user/__generated__/mutations.generated';
+export { default as UserProvider } from './user/UserProvider';
